refactor: new namings and remove unused codes
