+++ src/telemetry_forager/main.py
@@ -2,11 +2,11 @@ from typing import List
 from fastapi import FastAPI, Response
 import uvicorn
 from pydantic import BaseModel
-from healthstatus import HealthStatusClient, HealthReport
+from health import HealthStatusClient
 import jsons
 from datetime import datetime
 from config import AppConfig
-from model import ResourceParameter
+from model import ResourceParameter, ResourceHealthResult
 import log as Log
 from opentelemetry.trace import (
     SpanKind
@@ -29,69 +29,6 @@ class RequestResourceListParam(BaseModel):
 
 class RequestBodyParam(BaseModel):
     resources: List[RequestResourceListParam]
-
-class ResourceHealthAPIResult:
-
-    # availabilityState
-    # Available = 1, Unavailable = 0 or Unknown = 2
-    # converting into number is easier to style by Grafana "threshold" 
-
-    def __init__(self, 
-                 location='', 
-                 availabilityState='', 
-                 summary='', 
-                 reportedTime=None, 
-                 stateLastChangeTime=None) -> None:
-
-        self.location = location
-        self.availabilityState = availabilityState
-        self.summary = summary
-        self.reportedTime = (reportedTime if not None else datetime.now()).strftime("%B %d, %Y %H:%M:%S")
-        self.stateLastChangeTime = (stateLastChangeTime if not None else datetime.now()).strftime("%B %d, %Y %H:%M:%S")
-        self.displayText = ''
-
-        if availabilityState == 'Available':
-            self.availabilityState = 1
-            self.displayText = 'Available'
-        elif availabilityState == 'Unavailable':
-            self.availabilityState = 0
-            self.displayText = 'Unavailable'
-        else:
-            self.availabilityState = 2
-            self.displayText = 'Unknown'
-
-class RHResult:
-    def __init__(self, states: list[HealthReport]) -> None: #list[ResourceHealthAPIResult]) -> None:
-
-        # overallHealth
-        # Available = 1, Unavailable = 0, Partial = 2
-        # converting into number is easier to style by Grafana "threshold"
-
-        self.overallHealth = 1
-        self.overallSummary = ''
-
-        self.states: list[HealthReport] = states
-
-        self.set_overall_health_state()
-
-    def set_overall_health_state(self):
-
-        if not self.states:
-            return
-        
-        if all([x.availabilityState == 1 for x in self.states]):
-            self.overallHealth = 1
-            self.overallSummary = 'Available'
-            return
-        
-        if any([x.availabilityState == 0 for x in self.states]):
-            self.overallHealth = 0
-            self.overallSummary = 'Unavailable'
-            return
-        
-        if any([x.availabilityState == 2 for x in self.states]):
-            self.overallHealth = 2
-            self.overallSummary = 'Warning'
 
 
 def _get_subscription_id(resourceId: str) -> str:
@@ -132,7 +69,7 @@ def get_resource_params(rbp: RequestBodyParam) -> list[ResourceParameter]:
         
     return result
 
-def get_resource_health_states(resources: List[ResourceParameter]) -> RHResult:
+def get_resource_health_states(resources: List[ResourceParameter]) -> ResourceHealthResult:
 
     if not resources:
         return []
@@ -147,7 +84,7 @@ def get_resource_health_states(resources: List[ResourceParameter]) -> RHResult:
 
         healthStatuses.append(healthReport)
 
-    return RHResult(healthStatuses)
+    return ResourceHealthResult(healthStatuses)
